% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait.R
\name{wait}
\alias{wait}
\title{Function to wait for a specified number of seconds}
\usage{
wait(n_secs)
}
\arguments{
\item{n_secs}{Number of seconds to wait}
}
\description{
This function pauses the execution for a specified number of seconds.
}
\examples{
wait(5) # Waits for 5 seconds

}
